fix: estruturando as pastas e arquivos
