+++ lib/crypto_details/crypto_details_screen.dart
@@ -1,10 +1,10 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 
+import 'package:everest_card2_listagem/crypto_details/providers_details/providers_details.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'package:everest_card2_listagem/shared/api/model/crypto_view_data.dart';
-import 'package:everest_card2_listagem/portfolio/provider/crypto_provider.dart';
+import 'package:everest_card2_listagem/portfolio/model/crypto_view_data.dart';
 
 import '../portfolio/provider/isVisible_provider.dart';
 import 'widgets/body_crypto_details.dart';
