+++ lib/crypto_details/mapper/market_mapper.dart
@@ -1,5 +1,5 @@
 import '../model/market_graphic_view_data.dart';
-import '../response/market_graphic_response.dart';
+import '../../shared/api/response/market_graphic_response.dart';
 
 extension MarketMapper on MarketGraphicResponse {
   MarketGraphicViewData toViewMarkerGraphic() {
