+++ lib/crypto_details/model/market_graphic_view_data.dart
