+++ lib/portfolio/mapper/crypto_mapper.dart
@@ -1,5 +1,7 @@
+import 'package:decimal/decimal.dart';
+
 import '../model/crypto_view_data.dart';
-import '../response/crypto_results_response.dart';
+import '../../shared/api/response/crypto_results_response.dart';
 
 extension CryptoMapper on CryptoResultsResponse {
   List<CryptoViewData> toViewData() {
@@ -10,7 +12,7 @@ extension CryptoMapper on CryptoResultsResponse {
               symbol: result.symbol,
               name: result.name,
               image: result.image,
-              current_price: result.current_price,
+              current_price: Decimal.parse('${result.current_price}'),
               price_change_percentage_24h: result.price_change_percentage_24h),
         )
         .toList();
