+++ lib/portfolio/model/crypto_list_view_data.dart
@@ -1,5 +1,5 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:everest_card2_listagem/shared/api/model/crypto_view_data.dart';
+import 'package:everest_card2_listagem/portfolio/model/crypto_view_data.dart';
 
 class CryptoListViewData {
   List<CryptoViewData> cryptoListViewData;
