+++ lib/portfolio/useCase/total_crypto_value_use_case.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:decimal/decimal.dart';
-import 'package:everest_card2_listagem/shared/api/mapper/crypto_mapper.dart';
-import 'package:everest_card2_listagem/shared/api/model/crypto_view_data.dart';
+import 'package:everest_card2_listagem/portfolio/mapper/crypto_mapper.dart';
+import 'package:everest_card2_listagem/portfolio/model/crypto_view_data.dart';
 import 'package:everest_card2_listagem/shared/api/repository/crypto_repository.dart';
 
 class TotalCryptoValueUseCase {
