+++ lib/portfolio/view/home_portofolio_screen.dart
@@ -1,6 +1,5 @@
 import 'package:everest_card2_listagem/portfolio/widgtes/info_title_wallet_column_screen.dart';
 
-import '../../shared/total_wallet_model/total_wallet_model.dart';
 import '../widgtes/crypto_list_view.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
@@ -20,15 +19,13 @@ class _WalletScreenState extends ConsumerState<HomePortfolioScreen> {
 
   @override
   Widget build(BuildContext context) {
-    List<TotalWalletModel> totalWalletCrypto = [];
-
     return Scaffold(
       body: SafeArea(
         child: Column(
           children: [
             Padding(
               padding: const EdgeInsets.all(12.0),
-              child: InfoTitleColumnWallet(totalWalletModel: totalWalletCrypto),
+              child: InfoTitleColumnWallet(),
             ),
             const HomePortfolioColumn(),
           ],
