+++ lib/shared/routes/routes.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import '../../crypto_details/crypto_details_screen.dart';
 import '../../movimentations/view/movimentations_screen.dart';
 import '../../portfolio/view/home_portofolio_screen.dart';
-import '../api/model/crypto_view_data.dart';
+import '../../portfolio/model/crypto_view_data.dart';
 
 class Arguments {
   CryptoViewData cryptoData;
