--- lib/shared/total_wallet_model/total_wallet_model.dart
@@ -1,14 +1,0 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
-
-class TotalWalletModel {
-  String id;
-  double cryptoValueWallet;
-  TotalWalletModel({
-    required this.id,
-    required this.cryptoValueWallet,
-  });
-
-  @override
-  String toString() =>
-      'TotalWalletModel(id: $id,cryptoValueWallet: $cryptoValueWallet )';
-}
